back/quotes/serializers.py
from rest_framework import serializers
from .models import Quote

class QuoteSerializer(serializers.ModelSerializer):
    class Meta:
        model = Quote
        fields = ['id', 'text', 'author', 'date', 'created_at']
        read_only_fields = ['created_at']
