back/requirements.txt
django
gunicorn
django-heroku
djangorestframework
psycopg2-binary
django-cors-headers